Update development dependencies
So that we are using the latest bundler, rake, and rspec.

I also set the `.ruby-version` file so that developers would be using
the same version of ruby and updated the url to the active GitHub
repository.

Beyond that I removed thea badges from the README as they are outdated
and no longer configured correctly and I removed the code climate
configuration as it is outdated and no longer configured correctly.

This gets us back to the baseline of the library which is working great!